ec2-ssh/pyproject.toml
[build-system]
requires = ["setuptools", "wheel" , "bump-my-version"]
build-backend = "setuptools.build_meta"

[project]
name = "ec2_ssh"
description = "Command line utilities for finding and SSH-ing into your Amazon EC2 instances by tag (such as ‘Name’)."
authors = [{ name = "Modulr Finance" }]
license = {text = "No License"}
requires-python = ">=3.7"
dependencies = [
    "boto3~=1.34.49",
    "pick~=2.2.0",
    "PyYAML~=6.0.1",
    ]
dynamic = ["version"]

[project.scripts]
ec2-ssh = "ec2_ssh.cli:main"

[project.optional-dependencies]
deploy = [
    "bump-my-version==0.11.0",
    "twine<5.0.0",
]
test = [
    "black<24.0.0",
    "coverage<6.0",
    "pytest<8.0.0",
    "pytest-mock<4.0.0",
    "ruff==0.0.292",
]

[tool.setuptools.packages.find]
exclude = ["terraform*", "tests*", "docs*", "build"]

[tool.setuptools.dynamic]
version = {attr = "ec2_ssh.__version__"}

[tool.bumpversion]
current_version = "0.1.0"
commit = true
tag = true

[[tool.bumpversion.files]]
filename = "ec2_ssh/__init__.py"

[tool.ruff]
# Match black
line-length = 119
src = ["ec2_ssh", "tests"]

[tool.black]
line-length = 119
include = '\.pyi?$'
